Characterbase备份

+++ servers/src/application/entity/Card/CardAbstract/Character.go
@@ -3,7 +3,7 @@ package CardAbstract
 type Character interface {
 	Card
 	Attack(tempId int)
-	Hurt(tempId int, HurtHp int) //id是谁打的他
+	Hurt(tempId int, HurtHp float64) //id是谁打的他
 	Skill(tempId int)
 	Death(tempId int)
 }
