+++ servers/src/application/entity/Card/CardImpl/16.go
@@ -14,7 +14,7 @@ func NewCard16() *Card16 {
 func (c *Card16) Attack(tempId int) {
 
 }
-func (c *Card16) Hurt(tempId int, HurtHp int) {
+func (c *Card16) Hurt(tempId int, HurtHp float64) {
 }
 
 func (c *Card16) GetID() int {
