+++ servers/src/application/entity/Card/CardImpl/31.go
@@ -14,7 +14,7 @@ func NewCard31() *Card31 {
 func (c *Card31) Attack(tempId int) {
 
 }
-func (c *Card31) Hurt(tempId int, HurtHp int) {
+func (c *Card31) Hurt(tempId int, HurtHp float64) {
 }
 
 func (c *Card31) GetID() int {
