+++ servers/src/application/entity/Card/CardImpl/38.go
@@ -14,7 +14,7 @@ func NewCard38() *Card38 {
 func (c *Card38) Attack(tempId int) {
 
 }
-func (c *Card38) Hurt(tempId int, HurtHp int) {
+func (c *Card38) Hurt(tempId int, HurtHp float64) {
 }
 
 func (c *Card38) GetID() int {
