+++ servers/src/application/entity/Card/CardImpl/43.go
@@ -14,7 +14,7 @@ func NewCard43() *Card43 {
 func (c *Card43) Attack(tempId int) {
 
 }
-func (c *Card43) Hurt(tempId int, HurtHp int) {
+func (c *Card43) Hurt(tempId int, HurtHp float64) {
 }
 
 func (c *Card43) GetID() int {
